requirements.txt
aniso8601==1.2.0
appdirs==1.4.3
click==6.7
Flask==0.12
Flask-Cors==3.0.2
Flask-MySQL==1.4.0
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
PyMySQL==0.7.10
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
Werkzeug==0.12.1
MySQL-python==1.2.5
